requirements.txt
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
fastapi-camelcase==1.0.2
h11==0.12.0
httptools==0.1.1
loguru==0.5.3
passlib==1.7.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pyhumps==1.6.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
starlette==0.13.6
tabulate==0.8.9
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
yoyo-migrations==7.3.1
